Poisson Update

+++ aver_wait_time.cpp
@@ -10,11 +10,11 @@ static int s_TotCustNum = 0, s_TotWaitTime = 0;
 
 static queue<int> s_CustInLine;
 
-void cust_in(int InNum,int quenum) {
+void cust_in(int InNum, int quenum) {
 	int i;
 	if (quenum == 0) {
 		++s_TotCustNum;
-		g_AveWaitTime = ceil((double)s_TotWaitTime / s_TotCustNum);
+		g_AveWaitTime = (int)ceil((double)s_TotWaitTime / s_TotCustNum);
 	}
 	for (i = 0; i <= InNum - 1; ++i) {
 		s_CustInLine.push(g_Time);
@@ -26,6 +26,6 @@ void cust_out() {
 	++s_TotCustNum;
 	s_TotWaitTime += g_Time - s_CustInLine.back();
 	s_CustInLine.pop();
-	g_AveWaitTime = ceil((double)s_TotWaitTime / s_TotCustNum);
+	g_AveWaitTime = (int)ceil((double)s_TotWaitTime / s_TotCustNum);
 	return;
 }
